Minor package renaming in :settings:api
