+++ app/src/androidTest/java/com/javernaut/whatthecodec/screenshots/ScreenshotsTestSuite.kt
@@ -10,11 +10,11 @@ import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 import com.javernaut.whatthecodec.compose.theme.WhatTheCodecTheme
 import com.javernaut.whatthecodec.compose.theme.dynamic.ThemeViewModel
-import com.javernaut.whatthecodec.feature.settings.content.AudioStreamFeature
-import com.javernaut.whatthecodec.feature.settings.content.SubtitleStreamFeature
-import com.javernaut.whatthecodec.feature.settings.content.VideoStreamFeature
-import com.javernaut.whatthecodec.feature.settings.impl.content.completeEnumSet
-import com.javernaut.whatthecodec.feature.settings.theme.AppTheme
+import com.javernaut.whatthecodec.feature.settings.api.content.AudioStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.content.SubtitleStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.content.VideoStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.theme.AppTheme
+import com.javernaut.whatthecodec.feature.settings.data.content.completeEnumSet
 import com.javernaut.whatthecodec.home.presentation.model.ActualFrame
 import com.javernaut.whatthecodec.home.presentation.model.ActualPreview
 import com.javernaut.whatthecodec.home.presentation.model.AudioPage
