+++ app/src/main/java/com/javernaut/whatthecodec/compose/theme/dynamic/ThemeViewModel.kt
@@ -2,8 +2,8 @@ package com.javernaut.whatthecodec.compose.theme.dynamic
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.javernaut.whatthecodec.feature.settings.theme.AppTheme
-import com.javernaut.whatthecodec.feature.settings.theme.ThemeSettingsRepository
+import com.javernaut.whatthecodec.feature.settings.api.theme.AppTheme
+import com.javernaut.whatthecodec.feature.settings.api.theme.ThemeSettingsRepository
 import com.javernaut.whatthecodec.settings.presentation.stateIn
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.map
