+++ app/src/main/java/com/javernaut/whatthecodec/home/presentation/model/VideoPage.kt
@@ -1,8 +1,8 @@
 package com.javernaut.whatthecodec.home.presentation.model
 
-import com.javernaut.whatthecodec.feature.settings.content.AudioStreamFeature
-import com.javernaut.whatthecodec.feature.settings.content.SubtitleStreamFeature
-import com.javernaut.whatthecodec.feature.settings.content.VideoStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.content.AudioStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.content.SubtitleStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.content.VideoStreamFeature
 import io.github.javernaut.mediafile.AudioStream
 import io.github.javernaut.mediafile.SubtitleStream
 import io.github.javernaut.mediafile.VideoStream
