+++ app/src/main/java/com/javernaut/whatthecodec/settings/ui/SettingsScreen.kt
@@ -39,10 +39,10 @@ import com.javernaut.whatthecodec.compose.preference.PreferenceDivider
 import com.javernaut.whatthecodec.compose.preference.PreferenceGroup
 import com.javernaut.whatthecodec.compose.theme.WhatTheCodecTheme
 import com.javernaut.whatthecodec.compose.theme.dynamic.ThemeViewModel
-import com.javernaut.whatthecodec.feature.settings.content.AudioStreamFeature
-import com.javernaut.whatthecodec.feature.settings.content.SubtitleStreamFeature
-import com.javernaut.whatthecodec.feature.settings.content.VideoStreamFeature
-import com.javernaut.whatthecodec.feature.settings.theme.AppTheme
+import com.javernaut.whatthecodec.feature.settings.api.content.AudioStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.content.SubtitleStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.content.VideoStreamFeature
+import com.javernaut.whatthecodec.feature.settings.api.theme.AppTheme
 import com.javernaut.whatthecodec.home.ui.audio.displayableResource
 import com.javernaut.whatthecodec.home.ui.subtitle.displayableResource
 import com.javernaut.whatthecodec.home.ui.video.displayableResource
