+++ features/settings/api/src/main/kotlin/com/javernaut/whatthecodec/feature/settings/api/content/ContentSettingsRepository.kt
@@ -1,4 +1,4 @@
-package com.javernaut.whatthecodec.feature.settings.content
+package com.javernaut.whatthecodec.feature.settings.api.content
 
 import kotlinx.coroutines.flow.Flow
