+++ features/settings/api/src/main/kotlin/com/javernaut/whatthecodec/feature/settings/api/theme/AppTheme.kt
@@ -1,0 +1,7 @@
+package com.javernaut.whatthecodec.feature.settings.api.theme
+
+enum class AppTheme {
+    Light,
+    Dark,
+    Auto
+}
