+++ features/settings/api/src/main/kotlin/com/javernaut/whatthecodec/feature/settings/api/theme/ThemeSettings.kt
@@ -1,0 +1,5 @@
+package com.javernaut.whatthecodec.feature.settings.api.theme
+
+class ThemeSettings(
+    val selectedTheme: AppTheme
+)
