+++ features/settings/api/src/main/kotlin/com/javernaut/whatthecodec/feature/settings/api/theme/ThemeSettingsRepository.kt
@@ -1,4 +1,4 @@
-package com.javernaut.whatthecodec.feature.settings.theme
+package com.javernaut.whatthecodec.feature.settings.api.theme
 
 import kotlinx.coroutines.flow.Flow
