--- features/settings/api/src/main/kotlin/com/javernaut/whatthecodec/feature/settings/theme/AppTheme.kt
@@ -1,7 +1,0 @@
-package com.javernaut.whatthecodec.feature.settings.theme
-
-enum class AppTheme {
-    Light,
-    Dark,
-    Auto
-}
