--- features/settings/api/src/main/kotlin/com/javernaut/whatthecodec/feature/settings/theme/ThemeSettings.kt
@@ -1,5 +1,0 @@
-package com.javernaut.whatthecodec.feature.settings.theme
-
-class ThemeSettings(
-    val selectedTheme: AppTheme
-)
