+++ features/settings/data/src/main/kotlin/com/javernaut/whatthecodec/feature/settings/di/SettingsDataModule.kt
@@ -1,9 +1,9 @@
 package com.javernaut.whatthecodec.feature.settings.di
 
-import com.javernaut.whatthecodec.feature.settings.content.ContentSettingsRepository
-import com.javernaut.whatthecodec.feature.settings.impl.content.DefaultContentSettingsRepository
-import com.javernaut.whatthecodec.feature.settings.impl.theme.DefaultThemeSettingsRepository
-import com.javernaut.whatthecodec.feature.settings.theme.ThemeSettingsRepository
+import com.javernaut.whatthecodec.feature.settings.api.content.ContentSettingsRepository
+import com.javernaut.whatthecodec.feature.settings.api.theme.ThemeSettingsRepository
+import com.javernaut.whatthecodec.feature.settings.data.content.DefaultContentSettingsRepository
+import com.javernaut.whatthecodec.feature.settings.data.theme.DefaultThemeSettingsRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -11,7 +11,7 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-abstract class SettingsModule {
+abstract class SettingsDataModule {
 
     @Binds
     internal abstract fun bindThemeSettingsRepository(
